// apps/server/src/types.ts
export enum Role {
  ADMIN = "ADMIN",
  USER = "USER",
}

export enum VpnStatus {
  ONLINE = "ONLINE",
  OFFLINE = "OFFLINE",
  WARNING = "WARNING",
  PENDING = "PENDING",
}

export enum AuditAction {
  LOGIN = "LOGIN",
  LOGOUT = "LOGOUT",
  VPN_CREATE = "VPN_CREATE",
  VPN_DELETE = "VPN_DELETE",
  VPN_RESTART = "VPN_RESTART",
  TEMPLATE_EDIT = "TEMPLATE_EDIT",
  SUBSCRIPTION_STATUS = "SUBSCRIPTION_STATUS",
}